player component, forward ref, seekTo
